Unify sync and async paths.
