--- src/NCalc.Async/AsyncExpressionFunction.cs
@@ -1,3 +1,0 @@
-namespace NCalc;
-
-public delegate ValueTask<object?> AsyncExpressionFunction(AsyncExpressionFunctionData data);
