--- src/NCalc.Async/AsyncExpressionParameter.cs
@@ -1,3 +1,0 @@
-namespace NCalc;
-
-public delegate ValueTask<object?> AsyncExpressionParameter(AsyncExpressionParameterData data);
