--- src/NCalc.Async/AsyncExpressionParameterData.cs
@@ -1,8 +1,0 @@
-namespace NCalc;
-
-public class AsyncExpressionParameterData(Guid id, AsyncExpressionContext context, CancellationToken ct)
-{
-    public Guid Id { get; } = id;
-    public AsyncExpressionContext Context { get; } = context;
-    public CancellationToken CancellationToken { get; } = ct;
-}
