--- src/NCalc.Async/Factories/AsyncEvaluationVisitorFactory.cs
@@ -1,11 +1,0 @@
-using NCalc.Visitors;
-
-namespace NCalc.Factories;
-
-public sealed class AsyncEvaluationVisitorFactory : IAsyncEvaluationVisitorFactory
-{
-    public AsyncEvaluationVisitor Create(AsyncExpressionContext context)
-    {
-        return new AsyncEvaluationVisitor(context);
-    }
-}
