--- src/NCalc.Async/Factories/IAsyncEvaluationVisitorFactory.cs
@@ -1,8 +1,0 @@
-using NCalc.Visitors;
-
-namespace NCalc.Factories;
-
-public interface IAsyncEvaluationVisitorFactory
-{
-    public AsyncEvaluationVisitor Create(AsyncExpressionContext context);
-}
