--- src/NCalc.Async/Handlers/AsyncEvaluateFunctionHandler.cs
@@ -1,3 +1,0 @@
-namespace NCalc.Handlers;
-
-public delegate ValueTask AsyncEvaluateFunctionHandler(string name, AsyncFunctionArgs args);
