--- src/NCalc.Async/Handlers/AsyncEvaluateParameterHandler.cs
@@ -1,3 +1,0 @@
-namespace NCalc.Handlers;
-
-public delegate ValueTask AsyncEvaluateParameterHandler(string name, AsyncParameterArgs args);
