+++ src/NCalc.Async/README.md
@@ -1,6 +1,8 @@
 ﻿# NCalcAsync
 
-NCalc is a mathematical expression evaluator for .NET. 
-This assembly brings `async` support for NCalc.
+NCalc is a mathematical expression evaluator for .NET.
+`NCalcAsync` is obsolete and only redirects to `NCalc.Core`.
+
+Use `NCalc.Core` directly for the shared expression engine and async evaluation API.
 
 For more information, [read the docs](https://ncalc.github.io/ncalc/articles/async.html).
