+++ src/NCalc.Core/Expression.cs
@@ -1,0 +1,342 @@
+﻿using System.Diagnostics.CodeAnalysis;
+using NCalc.Cache;
+using NCalc.Domain;
+using NCalc.Exceptions;
+using NCalc.Factories;
+using NCalc.Handlers;
+using NCalc.Helpers;
+using NCalc.Visitors;
+
+namespace NCalc;
+
+/// <summary>
+/// This class represents a mathematical or logical expression that can be asynchronous evaluated.
+/// It supports caching, custom parameter and function evaluation, and options for handling null parameters and iterating over parameter collections.
+/// The class manages the parsing, validation, and evaluation of expressions, and provides mechanisms for error detection and reporting.
+/// </summary>
+public class Expression
+{
+    public IDictionary<string, ExpressionParameter> DynamicParameters
+    {
+        get => Context.DynamicParameters;
+        set => Context.DynamicParameters = value;
+    }
+
+    public IDictionary<string, ExpressionFunction> Functions
+    {
+        get => Context.Functions;
+        set => Context.Functions = value;
+    }
+
+    public IDictionary<string, AsyncExpressionFunction> AsyncFunctions
+    {
+        get => Context.AsyncFunctions;
+        set => Context.AsyncFunctions = value;
+    }
+
+    /// <summary>
+    /// Event triggered to handle function evaluation.
+    /// </summary>
+    public event EvaluateFunctionHandler EvaluateFunction
+    {
+        add => Context.EvaluateFunctionHandler += value;
+        remove => Context.EvaluateFunctionHandler -= value;
+    }
+
+    /// <summary>
+    /// Event triggered to handle async function evaluation.
+    /// </summary>
+    public event EvaluateAsyncFunctionHandler EvaluateAsyncFunction
+    {
+        add => Context.EvaluateAsyncFunctionHandler += value;
+        remove => Context.EvaluateAsyncFunctionHandler -= value;
+    }
+
+    /// <summary>
+    /// Event triggered to handle parameter evaluation.
+    /// </summary>
+    public event EvaluateParameterHandler EvaluateParameter
+    {
+        add => Context.EvaluateParameterHandler += value;
+        remove => Context.EvaluateParameterHandler -= value;
+    }
+
+    protected IEvaluationVisitorFactory EvaluationVisitorFactory { get; }
+    protected ExpressionContext Context { get; }
+
+    /// <summary>
+    /// Options for the expression evaluation.
+    /// </summary>
+    public ExpressionOptions Options
+    {
+        get => Context.Options;
+        set => Context.Options = value;
+    }
+
+    /// <summary>
+    /// Culture information for the expression evaluation.
+    /// </summary>
+    public CultureInfo CultureInfo
+    {
+        get => Context.CultureInfo;
+        set => Context.CultureInfo = value;
+    }
+
+    /// <summary>
+    /// Parameters for the expression evaluation.
+    /// </summary>
+    public IDictionary<string, object?> Parameters
+    {
+        get => Context.StaticParameters;
+        set => Context.StaticParameters = value;
+    }
+
+    /// <summary>
+    /// Textual representation of the expression.
+    /// </summary>
+    public string? ExpressionString { get; protected init; }
+
+    public LogicalExpression? LogicalExpression { get; set; }
+    public Exception? Error { get; private set; }
+    private ILogicalExpressionCache LogicalExpressionCache { get; }
+    private ILogicalExpressionFactory LogicalExpressionFactory { get; }
+
+    protected Expression(ExpressionContext? context = null)
+    {
+        LogicalExpressionCache = Cache.LogicalExpressionCache.GetInstance();
+        LogicalExpressionFactory = Factories.LogicalExpressionFactory.GetInstance();
+        Context = context ?? new ExpressionContext();
+        EvaluationVisitorFactory = new EvaluationVisitorFactory();
+    }
+
+    protected Expression(
+        string expressionString,
+        ExpressionContext context,
+        ILogicalExpressionFactory logicalExpressionFactory,
+        ILogicalExpressionCache logicalExpressionCache)
+    {
+        ExpressionString = expressionString;
+        LogicalExpressionCache = logicalExpressionCache;
+        LogicalExpressionFactory = logicalExpressionFactory;
+        Context = context;
+        EvaluationVisitorFactory = new EvaluationVisitorFactory();
+    }
+
+    protected Expression(
+        LogicalExpression logicalExpression,
+        ExpressionContext context,
+        ILogicalExpressionFactory logicalExpressionFactory,
+        ILogicalExpressionCache logicalExpressionCache)
+    {
+        LogicalExpression = logicalExpression;
+        LogicalExpressionCache = logicalExpressionCache;
+        LogicalExpressionFactory = logicalExpressionFactory;
+        EvaluationVisitorFactory = new EvaluationVisitorFactory();
+        Context = context;
+    }
+
+    public Expression(
+        string expression,
+        ExpressionContext context,
+        ILogicalExpressionFactory factory,
+        ILogicalExpressionCache cache,
+        IEvaluationVisitorFactory evaluationVisitorFactory) : this(expression, context, factory, cache)
+    {
+        EvaluationVisitorFactory = evaluationVisitorFactory;
+    }
+
+    public Expression(
+        LogicalExpression logicalExpression,
+        ExpressionContext context,
+        ILogicalExpressionFactory factory,
+        ILogicalExpressionCache cache,
+        IEvaluationVisitorFactory evaluationVisitorFactory) : this(logicalExpression, context, factory, cache)
+    {
+        EvaluationVisitorFactory = evaluationVisitorFactory;
+    }
+
+    public Expression(string? expression, ExpressionContext? context = null) : this(context)
+    {
+        ExpressionString = expression;
+    }
+
+    // ReSharper disable once RedundantOverload.Global
+    // Reason: False positive, ExpressionContext have implicit conversions.
+    public Expression(string? expression) : this(expression, ExpressionOptions.None)
+    {
+    }
+
+    public Expression(string? expression, ExpressionOptions options = ExpressionOptions.None,
+        CultureInfo? cultureInfo = null) : this(expression, new ExpressionContext(options, cultureInfo))
+    {
+    }
+
+    public Expression(LogicalExpression logicalExpression, ExpressionContext? context = null) : this(context)
+    {
+        LogicalExpression = logicalExpression ?? throw new
+            ArgumentException("Expression can't be null", nameof(logicalExpression));
+    }
+
+    // ReSharper disable once RedundantOverload.Global
+    // Reason: False positive, ExpressionContext have implicit conversions.
+    public Expression(LogicalExpression logicalExpression) : this(logicalExpression, ExpressionOptions.None)
+    {
+    }
+
+    public Expression(LogicalExpression logicalExpression, ExpressionOptions options = ExpressionOptions.None,
+        CultureInfo? cultureInfo = null) : this(logicalExpression, new ExpressionContext(options, cultureInfo))
+    {
+    }
+
+    /// <summary>
+    /// Evaluates the logical expression.
+    /// </summary>
+    /// <returns>The result of the evaluation.</returns>
+    /// <exception cref="NCalcException">Thrown when there is an error in the expression.</exception>
+    public object? Evaluate(CancellationToken ct = default)
+    {
+        var valueTask = EvaluateAsync(ct);
+
+        if (valueTask.IsCompletedSuccessfully)
+            return valueTask.Result;
+
+        return valueTask.AsTask().GetAwaiter().GetResult();
+    }
+
+    /// <summary>
+    /// Asynchronously evaluates the logical expression.
+    /// </summary>
+    /// <param name="ct">Cancellation token</param>
+    /// <returns>The result of the evaluation.</returns>
+    /// <exception cref="NCalcException">Thrown when there is an error in the expression.</exception>
+    public ValueTask<object?> EvaluateAsync(CancellationToken ct = default)
+    {
+        LogicalExpression ??= GetLogicalExpression(ct);
+
+        if (Error is not null)
+            throw Error;
+
+        if (Options.HasFlag(ExpressionOptions.AllowNullParameter))
+            Context.StaticParameters["null"] = null;
+
+        // If array evaluation, execute the same expression multiple times
+        if (Options.HasFlag(ExpressionOptions.IterateParameters))
+            return IterateParametersAsync(ct);
+
+        var evaluationVisitor = EvaluationVisitorFactory.Create(Context);
+
+        if (LogicalExpression is null)
+            return new ValueTask<object?>();
+
+        return LogicalExpression.Accept(evaluationVisitor, ct);
+    }
+
+    private async ValueTask<object?> IterateParametersAsync(CancellationToken ct)
+    {
+        var parameterEnumerators = ParametersHelper.GetEnumerators(Parameters, out var size);
+
+        var evaluationVisitor = EvaluationVisitorFactory.Create(Context);
+
+        if (LogicalExpression is null)
+            return null;
+
+        if (size == null)
+            return await LogicalExpression.Accept(evaluationVisitor, ct);
+
+        var results = new List<object?>();
+
+        for (var i = 0; i < size; i++)
+        {
+            foreach (var kvp in parameterEnumerators)
+            {
+                kvp.Value.MoveNext();
+                Parameters[kvp.Key] = kvp.Value.Current;
+            }
+
+            results.Add(await LogicalExpression.Accept(evaluationVisitor, ct));
+        }
+
+        return results;
+    }
+
+    /// <summary>
+    /// Returns a list with all parameter names from the expression.
+    /// </summary>
+    /// <param name="ct">Cancellation token</param>
+    public List<string> GetParameterNames(CancellationToken ct = default)
+    {
+        var parameterExtractionVisitor = new ParameterExtractionVisitor();
+        LogicalExpression ??= LogicalExpressionFactory.Create(ExpressionString!, CultureInfo, Context.Options, ct);
+        return LogicalExpression.Accept(parameterExtractionVisitor, ct);
+    }
+
+    /// <summary>
+    /// Returns a list with all function names from the expression.
+    /// </summary>
+    /// <param name="ct">Cancellation token</param>
+    public List<string> GetFunctionNames(CancellationToken ct = default)
+    {
+        var functionExtractionVisitor = new FunctionExtractionVisitor();
+        LogicalExpression ??= LogicalExpressionFactory.Create(ExpressionString!, CultureInfo, Context.Options, ct);
+        return LogicalExpression.Accept(functionExtractionVisitor, ct);
+    }
+
+    /// <summary>
+    /// Create the LogicalExpression in order to check syntax errors.
+    /// If errors are detected, the Error property contains the exception.
+    /// </summary>
+    /// <param name="ct">Cancellation token</param>
+    /// <returns>True if the expression syntax is correct, otherwise False.</returns>
+    [MemberNotNullWhen(true, nameof(Error))]
+    public bool HasErrors(CancellationToken ct = default)
+    {
+        try
+        {
+            Error = null;
+            LogicalExpression = LogicalExpressionFactory.Create(ExpressionString!, CultureInfo, Context.Options, ct);
+
+            // In case HasErrors() is called multiple times for the same expression
+            return LogicalExpression != null && Error != null;
+        }
+        catch (Exception exception)
+        {
+            Error = exception;
+            return true;
+        }
+    }
+
+    public LogicalExpression? GetLogicalExpression(CancellationToken ct = default)
+    {
+        if (string.IsNullOrEmpty(ExpressionString))
+        {
+            if (Options.HasFlag(ExpressionOptions.AllowNullOrEmptyExpressions))
+            {
+                return ExpressionString?.Length == 0 ? new ValueExpression(string.Empty) : null;
+            }
+
+            throw new NCalcException($"{nameof(ExpressionString)} cannot be null or empty.");
+        }
+
+        var isCacheEnabled = !Options.HasFlag(ExpressionOptions.NoCache);
+
+        LogicalExpression? logicalExpression = null;
+
+        if (isCacheEnabled && LogicalExpressionCache.TryGetValue(ExpressionString!, out logicalExpression))
+            return logicalExpression!;
+
+        try
+        {
+            Error = null;
+
+            logicalExpression = LogicalExpressionFactory.Create(ExpressionString!, CultureInfo, Context.Options, ct);
+            if (isCacheEnabled)
+                LogicalExpressionCache.Set(ExpressionString!, logicalExpression);
+        }
+        catch (Exception exception)
+        {
+            Error = exception;
+        }
+
+        return logicalExpression;
+    }
+}
