+++ src/NCalc.Core/Factories/EvaluationVisitorFactory.cs
