+++ src/NCalc.Core/Factories/ExpressionFactory.cs
