+++ src/NCalc.Core/Factories/IEvaluationVisitorFactory.cs
