+++ src/NCalc.Core/Factories/IExpressionFactory.cs
