+++ src/NCalc.Core/Handlers/AsyncExpressionFunction.cs
@@ -1,0 +1,3 @@
+namespace NCalc.Handlers;
+
+public delegate Task<object?> AsyncExpressionFunction(ExpressionFunctionData data);
