+++ src/NCalc.Core/Handlers/EvaluateAsyncFunctionHandler.cs
@@ -1,0 +1,3 @@
+namespace NCalc.Handlers;
+
+public delegate Task EvaluateAsyncFunctionHandler(string name, FunctionArgs args);
