+++ src/NCalc.Core/Handlers/EvaluateFunctionHandler.cs
