+++ src/NCalc.Core/Handlers/ExpressionFunction.cs
@@ -1,0 +1,3 @@
+namespace NCalc.Handlers;
+
+public delegate object? ExpressionFunction(ExpressionFunctionData data);
