+++ src/NCalc.Core/Handlers/ExpressionParameter.cs
@@ -1,3 +1,3 @@
-namespace NCalc;
+namespace NCalc.Handlers;
 
 public delegate object? ExpressionParameter(ExpressionParameterData data);
