+++ src/NCalc.Core/README.md
@@ -1,5 +1,7 @@
 ﻿# NCalc.Core
 
-Assembly with the core logic of NCalc. This is a base assembly for [NCalcSync](https://www.nuget.org/packages/NCalcSync) and [NCalcAsync](https://www.nuget.org/packages/NCalcAsync).
+Assembly with the core logic of NCalc. Use `NCalc.Core` directly for the shared `Expression` and `ExpressionContext` types, including async evaluation support.
+
+`NCalcSync` and `NCalcAsync` are obsolete redirect packages that point to `NCalc.Core`.
 
 For more information, [check our GitHub](https://www.github.com/NCalc/NCalc).
