+++ src/NCalc.DependencyInjection/NCalc.DependencyInjection.csproj
@@ -22,8 +22,7 @@
     </ItemGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\NCalc.Async\NCalc.Async.csproj" />
-      <ProjectReference Include="..\NCalc.Sync\NCalc.Sync.csproj" />
+      <ProjectReference Include="..\NCalc.Core\NCalc.Core.csproj" />
     </ItemGroup>
 
 </Project>
