--- src/NCalc.Sync/Expression.cs
@@ -1,155 +1,0 @@
-﻿using NCalc.Cache;
-using NCalc.Domain;
-using NCalc.Exceptions;
-using NCalc.Factories;
-using NCalc.Handlers;
-using NCalc.Helpers;
-
-namespace NCalc;
-
-/// <summary>
-/// This class represents a mathematical or logical expression that can be evaluated.
-/// It supports caching, custom parameter and function evaluation, and options for handling null parameters and iterating over parameter collections.
-/// The class manages the parsing, validation, and evaluation of expressions, and provides mechanisms for error detection and reporting.
-/// </summary>
-public class Expression : ExpressionBase<ExpressionContext>
-{
-    /// <summary>
-    /// Event triggered to handle function evaluation.
-    /// </summary>
-    public event EvaluateFunctionHandler EvaluateFunction
-    {
-        add => Context.EvaluateFunctionHandler += value;
-        remove => Context.EvaluateFunctionHandler -= value;
-    }
-
-    /// <summary>
-    /// Event triggered to handle parameter evaluation.
-    /// </summary>
-    public event EvaluateParameterHandler EvaluateParameter
-    {
-        add => Context.EvaluateParameterHandler += value;
-        remove => Context.EvaluateParameterHandler -= value;
-    }
-
-    public IDictionary<string, ExpressionParameter> DynamicParameters
-    {
-        get => Context.DynamicParameters;
-        set => Context.DynamicParameters = value;
-    }
-
-    public IDictionary<string, ExpressionFunction> Functions
-    {
-        get => Context.Functions;
-        set => Context.Functions = value;
-    }
-
-    protected IEvaluationVisitorFactory EvaluationVisitorFactory { get; }
-
-    private Expression(ExpressionContext? context = null) : base(context ?? new ExpressionContext())
-    {
-        EvaluationVisitorFactory = new EvaluationVisitorFactory();
-    }
-
-    public Expression(
-        string expression,
-        ExpressionContext context,
-        ILogicalExpressionFactory factory,
-        ILogicalExpressionCache cache,
-        IEvaluationVisitorFactory evaluationVisitorFactory) : base(expression, context, factory, cache)
-    {
-        EvaluationVisitorFactory = evaluationVisitorFactory;
-    }
-
-    public Expression(
-        LogicalExpression logicalExpression,
-        ExpressionContext context,
-        ILogicalExpressionFactory factory,
-        ILogicalExpressionCache cache,
-        IEvaluationVisitorFactory evaluationVisitorFactory) : base(logicalExpression, context, factory, cache)
-    {
-        EvaluationVisitorFactory = evaluationVisitorFactory;
-    }
-
-    public Expression(string? expression, ExpressionContext? context = null) : this(context)
-    {
-        ExpressionString = expression;
-    }
-
-    // ReSharper disable once RedundantOverload.Global
-    // Reason: False positive, ExpressionContext have implicit conversions.
-    public Expression(string? expression) : this(expression, ExpressionOptions.None)
-    {
-    }
-
-    public Expression(string? expression, ExpressionOptions options = ExpressionOptions.None,
-        CultureInfo? cultureInfo = null) : this(expression, new ExpressionContext(options, cultureInfo))
-    {
-    }
-
-    public Expression(LogicalExpression logicalExpression, ExpressionContext? context = null) : this(context)
-    {
-        LogicalExpression = logicalExpression ?? throw new
-            ArgumentException("Expression can't be null", nameof(logicalExpression));
-    }
-
-    // ReSharper disable once RedundantOverload.Global
-    // Reason: False positive, ExpressionContext have implicit conversions.
-    public Expression(LogicalExpression logicalExpression) : this(logicalExpression, ExpressionOptions.None)
-    {
-    }
-
-    public Expression(LogicalExpression logicalExpression, ExpressionOptions options = ExpressionOptions.None,
-        CultureInfo? cultureInfo = null) : this(logicalExpression, new ExpressionContext(options, cultureInfo))
-    {
-    }
-
-    /// <summary>
-    /// Evaluates the logical expression.
-    /// </summary>
-    /// <param name="ct">Cancellation token</param>
-    /// <returns>The result of the evaluation.</returns>
-    /// <exception cref="NCalcException">Thrown when there is an error in the expression.</exception>
-    public object? Evaluate(CancellationToken ct = default)
-    {
-        LogicalExpression ??= GetLogicalExpression(ct);
-
-        if (Error is not null)
-            throw Error;
-
-        if (Options.HasFlag(ExpressionOptions.AllowNullParameter))
-            Parameters["null"] = null;
-
-        // If array evaluation, execute the same expression multiple times
-        if (Options.HasFlag(ExpressionOptions.IterateParameters))
-            return IterateParameters(ct);
-
-        var evaluationVisitor = EvaluationVisitorFactory.Create(Context);
-        return LogicalExpression?.Accept(evaluationVisitor, ct);
-    }
-
-    private object? IterateParameters(CancellationToken ct)
-    {
-        var parameterEnumerators = ParametersHelper.GetEnumerators(Parameters, out var size);
-
-        var evaluationVisitor = EvaluationVisitorFactory.Create(Context);
-
-        if (size == null)
-            return LogicalExpression?.Accept(evaluationVisitor, ct);
-
-        var results = new List<object?>();
-
-        for (int i = 0; i < size; i++)
-        {
-            foreach (var kvp in parameterEnumerators)
-            {
-                kvp.Value.MoveNext();
-                Parameters[kvp.Key] = kvp.Value.Current;
-            }
-
-            results.Add(LogicalExpression?.Accept(evaluationVisitor, ct));
-        }
-
-        return results;
-    }
-}
