--- src/NCalc.Sync/ExpressionFunction.cs
@@ -1,3 +1,0 @@
-namespace NCalc;
-
-public delegate object? ExpressionFunction(ExpressionFunctionData args);
