+++ test/NCalc.Tests/NCalc.Tests.csproj
@@ -24,9 +24,8 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\src\NCalc.Async\NCalc.Async.csproj" />
+    <ProjectReference Include="..\..\src\NCalc.Core\NCalc.Core.csproj" />
     <ProjectReference Include="..\..\src\NCalc.DependencyInjection\NCalc.DependencyInjection.csproj" />
-    <ProjectReference Include="..\..\src\NCalc.Sync\NCalc.Sync.csproj" />
     <ProjectReference Include="..\..\src\Plugins\NCalc.Antlr\NCalc.Antlr.csproj" />
     <ProjectReference Include="..\..\src\Plugins\NCalc.LambdaCompilation\NCalc.LambdaCompilation.csproj" />
     <ProjectReference Include="..\..\src\Plugins\NCalc.MemoryCache\NCalc.MemoryCache.csproj" />
